cloudfunctions/requirements.txt
google-cloud-bigquery[pandas]==1.20.0
google-auth-oauthlib==0.4.0